Moved notification tests to rightful file
